Fix power usage calculation

+++ factorio-recipes/src/test/kotlin/BeNear.kt
@@ -9,9 +9,11 @@ import kotlin.math.roundToInt
 fun beNear(number: Double, epsilon: Double = 1e-6): Matcher<Number> = object : Matcher<Number> {
     override fun test(value: Number) = object : MatcherResult {
         override fun passed(): Boolean = abs(value.toDouble() - number) <= epsilon
-        override fun failureMessage(): String = "%f should be near %f (epsilon=%.1e)".format(value, number, epsilon)
+        override fun failureMessage(): String =
+            "%f should be near %f (epsilon=%.1e)".format(value.toDouble(), number, epsilon)
+
         override fun negatedFailureMessage(): String =
-            "%f should not be near %f (epsilon=%.1e)".format(value, number, epsilon)
+            "%f should not be near %f (epsilon=%.1e)".format(value.toDouble(), number, epsilon)
     }
 }
